requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
openpyxl==3.1.5
pandas==2.2.3
python-amazon-sp-api>=2.1.0
httpx==0.28.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
